pyproject.toml
[project]
name = "Extend_Pubmed_Network"
version = "0.1.0"
description = "Extend Pubmed Network"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "metapub>=0.5.12",
    "more-itertools>=10.6.0",
    "python-dotenv>=1.0.1",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
]
